Projects: Move ProxyModel class from Files
